작업
1. 데미지 텍스트 방식 변경
2. 정찰 시스템 구축 중

+++ Assets/3. Scripts/Charater/Enemy/EnemyAttack.cs
@@ -24,7 +24,7 @@ public class EnemyAttack : MonoBehaviour
 
     private bool IsAOEAttack = false;
     private Transform AOEexitPoint;         // 장판오브젝트 소환 포인트
-    public int AoeRadius;                   // 장판 반지름
+    public float AoeRadius;                   // 장판 반지름
     public int AoeDamage;                   // 장판오브젝트 데미지
     public int AoeTimes;                     // 장판의 공격 횟수
     public float AoeWaitforSeconds = 0f;    // 장판 공격간 쉬는 시간
